+++ src/warning/warning-printer.ts
@@ -52,6 +52,16 @@ export class WarningPrinter {
   async printWarning(warning: Warning) {
     const severity = this._severityToString(warning.severity);
     const range = warning.sourceRange;
+    if (!range) {
+      this._outStream.write(
+          `INTERNAL ERROR: Tried to print a '${warning.code}' ` +
+          `warning without a source range. Please report this!\n` +
+          `     https://github.com/Polymer/polymer-analyzer/issues/new\n`);
+      this._outStream.write(
+          `${this._severityToString(warning.severity)} ` +
+          `[${warning.code}] - ${warning.message}\n`);
+      return;
+    }
 
     if (this._options.verbosity === 'full') {
       this._outStream.write('\n\n');
@@ -77,9 +87,8 @@ export class WarningPrinter {
       default:
         const never: never = severity;
         throw new Error(
-            `Unknown severity value - ${
-                                        never
-                                      } - encountered while printing warning.`);
+            `Unknown severity value - ${never} - ` +
+            `encountered while printing warning.`);
     }
   }